Switch from numeric to INT type for nbr_distinct_users

+++ sql/moz-fx-data-shared-prod/google_play_store_derived/slow_startup_events_by_startup_type_v1/query.py
@@ -234,7 +234,7 @@ def main():
                 },
                 {
                     "name": "nbr_distinct_users",
-                    "type": "NUMERIC",
+                    "type": "INTEGER",
                     "mode": "NULLABLE",
                 },
             ],
